+++ sregistry/main/aws/__init__.py
@@ -1,6 +1,6 @@
 '''
 
-Copyright (C) 2017-2018 Vanessa Sochat.
+Copyright (C) 2018 Vanessa Sochat.
 
 This program is free software: you can redistribute it and/or modify it
 under the terms of the GNU Affero General Public License as published by
@@ -26,45 +26,31 @@ import os
 
 # The core of nvidia is actually docker
 from sregistry.main.docker.api import ( 
-    create_metadata_tar, 
-    download_layers, 
-    get_manifest_selfLink,
     get_config, 
-    get_digests,
     get_environment_tar, 
-    get_layer,
-    get_layerLink,
-    get_manifest,
-    get_manifests,
     get_download_cache,
-    get_size,
-    extract_env,
-    extract_labels,
-    extract_runscript,
-    update_token
+    get_size
 )
-
-from sregistry.main.docker.pull import _pull
-from .pull import pull
+from sregistry.main.aws.api import ( download_layers, update_token, get_digests,
+                                     get_manifest )
+from .pull import ( pull, _pull )
 
 class Client(ApiConnection):
 
     def __init__(self, secrets=None, **kwargs):
-        '''to work with nvidia container registry, we do the following:
+        '''to work with nvidia container registry. If you need help with
+           their API, the best reference is:
 
-        1. set the base to index.docker.io, or defined in environment 
-        2. assume starting with v2 schema, we won't reverse layers 
-        3. set headers to ask for version 2 schema first
-        4. update secrets, 1st priority environment, then .docker/config.json 
-        5. update headers based on secrets
+           https://docs.aws.amazon.com/AmazonECR/latest/APIReference/ecr-api.pdf
  
         '''
-        self._set_base()
         self.reverseLayers = False
         self._reset_headers()
         self._update_secrets()
+        self._set_base()
         self._update_headers()
         super(ApiConnection, self).__init__(**kwargs)
+
 
     def _speak(self):
         '''if you want to add an extra print (of a parameter, for example)
@@ -87,29 +73,29 @@ class Client(ApiConnection):
         accept += 'application/json'
 
         self.headers = {"Accept": accept,
-                        'Content-Type': 'application/json; charset=utf-8'}
+                        "Accept-Encoding": "identity",
+                        'Content-Type': 'application/x-amz-json-1.1; application/json; charset=utf-8'}
 
-    def _set_base(self):
+    def _set_base(self, zone=None):
         '''set the API base or default to use Docker Hub. The user is able
            to set the base, api version, and protocol via a settings file
            of environment variables:
- 
-           SREGISTRY_NVIDIA_BASE: defaults to nvcr.io
-           SREGISTRY_NVIDIA_TOKEN: defaults to $oauthtoken
-           SREGISTRY_NVIDIA_VERSION: defaults to v2
-           SREGISTRY_NVIDIA_NO_HTTPS: defaults to not set (so https)
-
         '''
-        base = self._get_setting('SREGISTRY_NVIDIA_BASE')
-        version = self._get_setting('SREGISTRY_NVIDIA_VERSION')
+        if hasattr(self.aws._client_config, 'region_name'):
+            zone = self.aws._client_config.region_name
+
+        aws_id = self._get_setting('SREGISTRY_AWS_ID')
+        aws_zone = self._get_setting('SREGISTRY_AWS_ZONE', zone)
+        base = self._get_setting('SREGISTRY_AWS_BASE')
+        version = self._get_setting('SREGISTRY_AWS_VERSION', 'v2')
+
+        if aws_id is None or aws_zone is None:
+            bot.exit('Please export SREGISTRY_AWS_ID and SREGISTRY_AWS_ZONE.')
 
         if base is None:
-            base = "nvcr.io"
+            base = "%s.dkr.ecr.%s.amazonaws.com" % (aws_id, aws_zone)
 
-        if version is None:
-            version = "v2"
-
-        nohttps = self._get_setting('SREGISTRY_NVIDIA_NOHTTPS')
+        nohttps = self._get_setting('SREGISTRY_AWS_NOHTTPS')
         if nohttps is None:
             nohttps = "https://"
         else:
@@ -117,7 +103,6 @@ class Client(ApiConnection):
 
         # <protocol>://<base>/<version>
         self.base = "%s%s/%s" %(nohttps, base.strip('/'), version)
-
 
     def _update_secrets(self):
         '''update secrets will take a secrets credential file
@@ -127,21 +112,15 @@ class Client(ApiConnection):
            using Docker Hub, if we find a .docker secrets file, we update
            from there.
         '''
-        # If the user has defined secrets, use them
-        token = self._get_and_update_setting('SREGISTRY_NVIDIA_TOKEN')
-        username = self._get_and_update_setting('SREGISTRY_NVIDIA_USERNAME')
+        bot.debug('Creating aws client...')
+        try:
+            from awscli.clidriver import create_clidriver
+        except:
+            bot.exit('Please install pip install sregistry[aws]')
 
-        if username is None:
-            username = "$oauthtoken"
-
-        if token is None:
-            bot.error('You must have an Nvidia GPU cloud token to use it.')
-            sys.exit(1)
-
-        # Option 1: the user exports username and password
-        if token is not None:
-            auth = basic_auth_header(username, token)
-            self.headers.update(auth)
+        driver = create_clidriver()
+        self.aws = driver.session.create_client('ecr')
+        
 
     def __str__(self):
         return type(self)
@@ -152,19 +131,10 @@ Client.pull = pull
 Client._pull = _pull
 
 # Api functions for image layers and manifests (hidden)
-Client._create_metadata_tar = create_metadata_tar
 Client._download_layers = download_layers
-Client._extract_runscript = extract_runscript
-Client._extract_labels = extract_labels
-Client._extract_env = extract_env
-Client._get_config = get_config
 Client._get_digests = get_digests
 Client._get_download_cache = get_download_cache
-Client._get_layer = get_layer
-Client._get_layerLink = get_layerLink
 Client._get_manifest = get_manifest
-Client._get_manifests = get_manifests
 Client._get_size = get_size
 Client._update_token = update_token
-Client._get_manifest_selfLink = get_manifest_selfLink
 Client._get_environment_tar = get_environment_tar