+++ sregistry/main/workers/worker.py
@@ -2,8 +2,6 @@
 
 sregistry.api: base template for making a connection to an API
 
-Copyright (C) 2017-2018 The Board of Trustees of the Leland Stanford Junior
-University.
 Copyright (C) 2017-2018 Vanessa Sochat.
 
 This program is free software: you can redistribute it and/or modify it